requirements.txt
flask
playwright
botbuilder-core
botbuilder-schema
aiohttp
dotenv